requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.5.0
pyyaml>=6.0
pulp>=2.7.0
plotly>=5.18.0
requests>=2.31.0
python-dateutil>=2.8.2
joblib>=1.3.0
scipy>=1.11.0